requirements.txt
Pillow
piexif
pikepdf
python-docx
